🖼🦄 finished refactor
