+++ package.json
@@ -22,7 +22,8 @@
     "pug": "^3.0.2",
     "passport": "^0.4.1",
     "express-session": "^1.17.2",
-    "passport-local": "^1.0.0"
+    "passport-local": "^1.0.0",
+    "bcrypt": "^5.0.1"
   },
   "repository": {
     "type": "git",
