+++ server.js
@@ -4,11 +4,10 @@ const express = require("express");
 const myDB = require("./connection");
 const fccTesting = require("./freeCodeCamp/fcctesting.js");
 const pug = require("pug");
-const session = require("express-session");
-const passport = require("passport");
-const LocalStrategy = require("passport-local");
 const mongodb = require("mongodb");
-const ObjectID = require("mongodb").ObjectID;
+
+const auth = require("./auth");
+const routes = require("./routes");
 
 const app = express();
 
@@ -17,130 +16,14 @@ app.use("/public", express.static(process.cwd() + "/public"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-app.use(
-  session({
-    secret: process.env.SESSION_SECRET,
-    resave: true,
-    saveUninitialized: true,
-    cookie: { secure: false }
-  })
-);
-
-app.use(passport.initialize());
-app.use(passport.session());
-
 app.set("view engine", "pug");
 app.set("views/", "./views");
-
-function ensureAuthenticated(req, res, next) {
-  if (req.isAuthenticated()) {
-    return next();
-  }
-  res.redirect("/");
-}
 
 myDB(async client => {
   const myDataBase = await client.db("advancedNode2021").collection("users");
 
-  // Be sure to change the title
-  app.route("/").get((req, res) => {
-    // Change the response to render the Pug template
-    res.render("pug", {
-      title: "Connected to Database",
-      message: "Please login",
-      showLogin: true,
-      showRegistration: true
-    });
-  });
-
-  app
-    .route("/login")
-    .post(
-      passport.authenticate("local", { failureRedirect: "/" }),
-      (req, res) => {
-        res.redirect("/profile");
-      }
-    );
-
-  app.route("/profile").get(ensureAuthenticated, (req, res) =>
-    res.render("/pug/profile", {
-      username: req.user.username
-    })
-  );
-
-  app.route("/logout").get((req, res) => {
-    req.logout();
-    res.redirect("/");
-  });
-
-  app.route("/register").post(
-    (req, res, next) => {
-      myDataBase.findOne({ username: req.body.username }, function(err, user) {
-        if (err) {
-          next(err);
-        } else if (user) {
-          res.redirect("/");
-        } else {
-          myDataBase.insertOne(
-            {
-              username: req.body.username,
-              password: req.body.password
-            },
-            (err, doc) => {
-              if (err) {
-                res.redirect("/");
-              } else {
-                // minor modification for mongoDB v4
-                myDataBase.findOne({ _id: doc.insertedId }, function(err, doc) {
-                  if (err) next(err);
-                  next(null, doc);
-                });
-              }
-            }
-          );
-        }
-      });
-    },
-    passport.authenticate("local", { failureRedirect: "/" }),
-    (req, res, next) => {
-      res.redirect("/profile");
-    }
-  );
-
-  app.use((req, res, next) => {
-    res
-      .status(404)
-      .type("text")
-      .send("Not Found");
-  });
-
-  // Serialization and deserialization here...
-  passport.serializeUser((user, done) => {
-    done(null, user._id);
-  });
-  passport.deserializeUser((id, done) => {
-    myDataBase.findOne({ _id: new ObjectID(id) }, (err, doc) => {
-      done(null, doc);
-    });
-  });
-
-  passport.use(
-    new LocalStrategy(function(username, password, done) {
-      myDataBase.findOne({ username: username }, function(err, user) {
-        console.log("User " + username + " attempted to log in.");
-        if (err) {
-          return done(err);
-        }
-        if (!user) {
-          return done(null, false);
-        }
-        if (password !== user.password) {
-          return done(null, false);
-        }
-        return done(null, user);
-      });
-    })
-  );
+  auth(app, myDataBase);
+  routes(app, myDataBase);
 }).catch(e => {
   app.route("/").get((req, res) => {
     res.render("pug", { title: e, message: "Unable to login" });
